refactor background and buffer headers
