+++ ENIGMAsystem/SHELL/Graphics_Systems/Direct3D11/DX11background.cpp
@@ -23,7 +23,7 @@
 #include "Graphics_Systems/General/GSbackground.h"
 #include "Graphics_Systems/General/GStextures.h"
 #include "Universal_System/nlpo2.h"
-#include "Universal_System/backgroundstruct.h"
+#include "Universal_System/background_internal.h"
 #include "Universal_System/spritestruct.h"
 
 #ifdef DEBUG_MODE
