+++ ENIGMAsystem/SHELL/Graphics_Systems/Direct3D11/DX11tiles.cpp
@@ -38,7 +38,7 @@
 #include "Universal_System/depth_draw.h"
 #include <algorithm>
 #include "../General/GSbackground.h"
-#include "Universal_System/backgroundstruct.h"
+#include "Universal_System/background_internal.h"
 #include "../General/GStextures.h"
 
 #include "Direct3D11Headers.h"
