+++ ENIGMAsystem/SHELL/Graphics_Systems/Direct3D9/DX9textures.cpp
@@ -23,7 +23,7 @@
 #include "../General/GStextures.h"
 #include "DX9TextureStruct.h"
 #include "Universal_System/image_formats.h"
-#include "Universal_System/backgroundstruct.h"
+#include "Universal_System/background_internal.h"
 #include "Universal_System/spritestruct.h"
 #include "Graphics_Systems/graphics_mandatory.h"
