+++ ENIGMAsystem/SHELL/Graphics_Systems/None/NONEscreen.cpp
@@ -28,7 +28,7 @@
 using namespace std;
 
 #include "Universal_System/image_formats.h"
-#include "Universal_System/backgroundstruct.h"
+#include "Universal_System/background_internal.h"
 #include "Universal_System/var4.h"
 #include "Universal_System/estring.h"
