+++ ENIGMAsystem/SHELL/Graphics_Systems/None/NONEtiles.cpp
@@ -20,7 +20,7 @@
 #include "Universal_System/depth_draw.h"
 #include <algorithm>
 #include "../General/GSbackground.h"
-#include "Universal_System/backgroundstruct.h"
+#include "Universal_System/background_internal.h"
 #include "../General/GStextures.h"
 #include "../General/GStiles.h"
 #include "../General/GLtilestruct.h"
