+++ ENIGMAsystem/SHELL/Graphics_Systems/OpenGL1/GLtiles.cpp
@@ -20,31 +20,13 @@
 #include "Universal_System/depth_draw.h"
 #include <algorithm>
 #include "../General/GSbackground.h"
-#include "Universal_System/backgroundstruct.h"
+#include "Universal_System/background.h"
+#include "Universal_System/background_internal.h"
 #include "../General/GStextures.h"
 #include "../General/GStiles.h"
 #include "../General/GLtilestruct.h"
 #include "../General/OpenGLHeaders.h"
 #include "GLTextureStruct.h"
-
-#ifdef DEBUG_MODE
-  #include <string>
-  #include "libEGMstd.h"
-  #include "Widget_Systems/widgets_mandatory.h"
-  #define get_background(bck2d,back)\
-    if (back < 0 or size_t(back) >= enigma::background_idmax or !enigma::backgroundstructarray[back]) {\
-      show_error("Attempting to draw non-existing background " + toString(back), false);\
-      return;\
-    }\
-    const enigma::background *const bck2d = enigma::backgroundstructarray[back];
-#else
-  #define get_background(bck2d,back)\
-    const enigma::background *const bck2d = enigma::backgroundstructarray[back];
-#endif
-
-#define __GETR(x) ((x & 0x0000FF))
-#define __GETG(x) ((x & 0x00FF00) >> 8)
-#define __GETB(x) ((x & 0xFF0000) >> 16)
 
 namespace enigma
 {
