+++ ENIGMAsystem/SHELL/Graphics_Systems/OpenGL3/GL3background.cpp
@@ -22,7 +22,7 @@
 
 #include "Universal_System/image_formats.h"
 #include "Universal_System/nlpo2.h"
-#include "Universal_System/backgroundstruct.h"
+#include "Universal_System/background_internal.h"
 #include "Graphics_Systems/graphics_mandatory.h"
 #include "Universal_System/spritestruct.h"
