+++ ENIGMAsystem/SHELL/Graphics_Systems/OpenGL3/GL3surface.cpp
@@ -31,7 +31,7 @@ using namespace std;
 #include <stdio.h> //for file writing (surface_save)
 #include "Universal_System/nlpo2.h"
 #include "Universal_System/spritestruct.h"
-#include "Universal_System/backgroundstruct.h"
+#include "Universal_System/background_internal.h"
 #include "Collision_Systems/collision_types.h"
 
 #define __GETR(x) (gs_scalar)(((x & 0x0000FF))/255.0)
