+++ ENIGMAsystem/SHELL/Universal_System/texture_atlas.cpp
@@ -21,7 +21,7 @@
 
 #include "Universal_System/spritestruct.h"
 #include "Universal_System/fontstruct.h"
-#include "Universal_System/backgroundstruct.h"
+#include "Universal_System/background_internal.h"
 
 #include "Universal_System/nlpo2.h"
 #include "Graphics_Systems/graphics_mandatory.h"
